+++ src/main/java/dev/willyelton/crystal_tools/common/compat/jei/CrystalToolsCraftingCategoryExtension.java
@@ -14,6 +14,7 @@ import java.util.List;
 import java.util.Optional;
 
 public class CrystalToolsCraftingCategoryExtension implements ICraftingCategoryExtension<CrystalToolsRecipe> {
+
     @Override
     public Optional<ResourceLocation> getRegistryName(RecipeHolder<CrystalToolsRecipe> recipe) {
         return Optional.of(recipe.id());