+++ src/main/java/dev/willyelton/crystal_tools/common/events/RegisterDataMapsEvent.java
@@ -13,5 +13,6 @@ public class RegisterDataMapsEvent {
         event.register(DataMaps.GENERATOR_GEMS);
         event.register(DataMaps.GENERATOR_METALS);
         event.register(DataMaps.MOB_SKULLS);
+        event.register(DataMaps.PEDESTAL_ACTIONS);
     }
 }